slash commands and more!

+++ Commands/moderation/purge.js
@@ -2,6 +2,7 @@
 const makeEmbed = require('../../functions/embed');
 const sendAndDelete = require("../../functions/sendAndDelete");
 const Command = require("../../Classes/Command");
+const colors = require("../../config/colors.json");
 
 let purge = new Command("purge");
 
@@ -17,12 +18,32 @@ purge.set({
 	requiredPerms   : "MANAGE_MESSAGES",
 	worksInDMs      : false,
 	isDevOnly       : false,
-	isSlashCommand  : false
+	isSlashCommand  : true,
+    options			: [
+        {
+            name : "amount",
+            description : "The amount of new message to delete (max = 100)",
+            required : true,
+            type: 4,
+		},
+		
+	],
 });
 
-purge.execute = function(message, args, server) {
+purge.execute = function(message, args, server,isSlash) {
 
-    let number = parseInt(args[0]);
+    const modLog = message.guild.channels.cache.get(server.logs.warningLog);
+    let author;
+    let number;
+    if(isSlash){
+        author = message.user;
+        number = args[0].value;
+    } else{
+        author = message.author;
+        number = parseInt(args[0]);
+    }
+
+
     if(isNaN(number)){
         const embed = makeEmbed('invalid number',this.usage, server);
         sendAndDelete(message,embed, server);
@@ -36,13 +57,21 @@ purge.execute = function(message, args, server) {
         sendAndDelete(message,embed, server);
         return false;
     } else{ 
+        
         message.channel.bulkDelete(number, true)
         .then(m=>{
-            message.channel.send(`Done! ✅`)
-            .then(msg=> setTimeout(() => {
-                msg.delete().catch(e=>e);
-            }, 4000))
+            if(isSlash){
+                sendAndDelete(message, "Done! ✅", server);
+            } else{
+                message.channel.send(`Done! ✅`)
+                .then(msg=> setTimeout(() => {
+                    msg.delete().catch(e=>e);
+                }, 4000))
+            }
+            const logEmbed = makeEmbed("Message purge",`The user <@${author.id}>[${author.id}] has purged ${number} message from <#${message.channel.id}> `,"ff7f00",true);
+            if(modLog)modLog.send({embeds:[logEmbed]});
         })
+
         .catch(console.error);
         return true;
     }
